refactor: consolidate parser into single file
Merge 20 separate parse_*.c/h files and parser_util.* into a unified
parser.c (1864 lines) with logical organization: utilities, types,
expressions, statements, declarations, imports, and entry points.
