--- w0/parse_block.c
@@ -1,24 +1,0 @@
-#include "parse_block.h"
-
-#include "parse_statement.h"
-#include "parser_util.h"
-
-Node* parse_block(Parser* parser) {
-    Node* block = node_new(NODE_BLOCK, parser->previous.line, parser->previous.column);
-    if (!block) {
-        parse_error(parser, "Out of memory");
-        return NULL;
-    }
-    nodelist_init(&block->as.block.stmts);
-
-    while (!check_token(parser, TOK_RBRACE) && !check_token(parser, TOK_EOF)) {
-        Node* stmt = parse_statement(parser);
-        if (stmt)
-            nodelist_push(&block->as.block.stmts, stmt);
-        if (parser->panic_mode)
-            synchronize(parser);
-    }
-
-    consume_token(parser, TOK_RBRACE, "Expected '}' after block");
-    return block;
-}
