--- w0/parse_block.h
@@ -1,9 +1,0 @@
-#ifndef WHIST_PARSE_BLOCK_H
-#define WHIST_PARSE_BLOCK_H
-
-#include "ast.h"
-#include "parser.h"
-
-Node* parse_block(Parser* parser);
-
-#endif
