--- w0/parse_defer_stmt.h
@@ -1,9 +1,0 @@
-#ifndef WHIST_PARSE_DEFER_STMT_H
-#define WHIST_PARSE_DEFER_STMT_H
-
-#include "ast.h"
-#include "parser.h"
-
-Node* parse_defer_stmt(Parser* parser);
-
-#endif
