--- w0/parse_expression.h
@@ -1,9 +1,0 @@
-#ifndef WHIST_PARSE_EXPRESSION_H
-#define WHIST_PARSE_EXPRESSION_H
-
-#include "ast.h"
-#include "parser.h"
-
-Node* parse_expression(Parser* parser);
-
-#endif
