--- w0/parse_extern_decls.h
@@ -1,9 +1,0 @@
-#ifndef WHIST_PARSE_EXTERN_DECLS_H
-#define WHIST_PARSE_EXTERN_DECLS_H
-
-#include "ast.h"
-#include "parser.h"
-
-Node* parse_extern_decls(Parser* parser, int is_public);
-
-#endif
