--- w0/parse_for_stmt.h
@@ -1,9 +1,0 @@
-#ifndef WHIST_PARSE_FOR_STMT_H
-#define WHIST_PARSE_FOR_STMT_H
-
-#include "ast.h"
-#include "parser.h"
-
-Node* parse_for_stmt(Parser* parser);
-
-#endif
