--- w0/parse_foreach_stmt.h
@@ -1,9 +1,0 @@
-#ifndef WHIST_PARSE_FOREACH_STMT_H
-#define WHIST_PARSE_FOREACH_STMT_H
-
-#include "ast.h"
-#include "parser.h"
-
-Node* parse_foreach_stmt(Parser* parser);
-
-#endif
