--- w0/parse_func_decl.c
@@ -1,133 +1,0 @@
-#include <stdlib.h>
-#include <string.h>
-
-#include "parse_enum_decl.h"
-#include "parser.h"
-#include "parser_util.h"
-
-Node* parse_type(Parser* parser);
-Node* parse_block(Parser* parser);
-
-// Declaration parsing
-Node* parse_func_decl(Parser* parser, int is_public) {
-    // Check for method receiver: func (Type) or func (const Type)
-    char* receiver_type     = NULL;
-    int   receiver_type_len = 0;
-    int   receiver_is_const = 0;
-
-    if (check_token(parser, TOK_LPAREN)) {
-        advance_token(parser); // consume '('
-
-        // Check for 'const' modifier
-        if (match_token(parser, TOK_CONST)) {
-            receiver_is_const = 1;
-        }
-
-        // Expect struct type name
-        Token recv_type = parser->current;
-        consume_token(parser, TOK_IDENT, "Expected receiver type name");
-        receiver_type = copy_token_string(&recv_type);
-        if (!receiver_type) {
-            parse_error(parser, "Out of memory");
-            return NULL;
-        }
-        receiver_type_len = recv_type.length;
-
-        consume_token(parser, TOK_RPAREN, "Expected ')' after receiver type");
-    }
-
-    Token name = parser->current;
-    consume_token(parser, TOK_IDENT, "Expected function name");
-
-    Node* node = node_new(NODE_FUNC_DECL, name.line, name.column);
-    if (!node) {
-        free(receiver_type);
-        parse_error(parser, "Out of memory");
-        return NULL;
-    }
-    func_decl_node* fdn = &node->as.func_decl;
-
-    fdn->is_public         = is_public;
-    fdn->is_extern         = 0;
-    fdn->receiver_type     = receiver_type;
-    fdn->receiver_type_len = receiver_type_len;
-    fdn->receiver_is_const = receiver_is_const;
-    fdn->name              = copy_token_string(&name);
-    if (!fdn->name) {
-        parse_error(parser, "Out of memory");
-        return NULL;
-    }
-    fdn->name_length = name.length;
-    nodelist_init(&fdn->params);
-
-    consume_token(parser, TOK_LPAREN, "Expected '(' after function name");
-
-    // Parameters
-    if (!check_token(parser, TOK_RPAREN)) {
-        do {
-            // Check for 'const' modifier
-            int param_is_const = 0;
-            if (match_token(parser, TOK_CONST)) {
-                param_is_const = 1;
-            }
-
-            Token param_name = parser->current;
-            consume_token(parser, TOK_IDENT, "Expected parameter name");
-
-            Node* param = node_new(NODE_PARAM, param_name.line, param_name.column);
-            if (!param) {
-                parse_error(parser, "Out of memory");
-                return NULL;
-            }
-            param->as.param.name = copy_token_string(&param_name);
-            if (!param->as.param.name) {
-                parse_error(parser, "Out of memory");
-                return NULL;
-            }
-            param->as.param.name_length = param_name.length;
-            param->as.param.type        = NULL;
-            param->as.param.is_const    = param_is_const;
-
-            if (match_token(parser, TOK_COLON)) {
-                param->as.param.type = parse_type(parser);
-            }
-
-            nodelist_push(&fdn->params, param);
-        } while (match_token(parser, TOK_COMMA));
-    }
-
-    consume_token(parser, TOK_RPAREN, "Expected ')' after parameters");
-
-    // Return type
-    fdn->return_type = NULL;
-    if (match_token(parser, TOK_COLON)) {
-        fdn->return_type = parse_type(parser);
-    }
-
-    // Body
-    if (check_token(parser, TOK_LBRACE) == 0) {
-        consume_token(parser, TOK_SEMICOLON, "Expected ';' after function declaration");
-        // extern function declaration
-        fdn->body = NULL;
-        return node;
-    }
-
-    consume_token(parser, TOK_LBRACE, "Expected '{' before function body");
-    fdn->body = parse_block(parser);
-
-    // Copy current file's direct imports to accessible_modules
-    // This determines which library modules this function can access
-    fdn->accessible_modules_count = parser->direct_imports_count;
-    if (parser->direct_imports_count > 0) {
-        fdn->accessible_modules = malloc(parser->direct_imports_count * sizeof(char*));
-        if (fdn->accessible_modules) {
-            for (int i = 0; i < parser->direct_imports_count; i++) {
-                fdn->accessible_modules[i] = strdup(parser->direct_imports[i]);
-            }
-        }
-    } else {
-        fdn->accessible_modules = NULL;
-    }
-
-    return node;
-}
