--- w0/parse_func_decl.h
@@ -1,9 +1,0 @@
-#ifndef WHIST_PARSE_FUNC_DECL_H
-#define WHIST_PARSE_FUNC_DECL_H
-
-#include "ast.h"
-#include "parser.h"
-
-Node* parse_func_decl(Parser* parser, int is_public);
-
-#endif
