--- w0/parse_if_stmt.h
@@ -1,9 +1,0 @@
-#ifndef WHIST_PARSE_IF_STMT_H
-#define WHIST_PARSE_IF_STMT_H
-
-#include "ast.h"
-#include "parser.h"
-
-Node* parse_if_stmt(Parser* parser);
-
-#endif
