--- w0/parse_primary.h
@@ -1,9 +1,0 @@
-#ifndef WHIST_PARSE_PRIMARY_H
-#define WHIST_PARSE_PRIMARY_H
-
-#include "ast.h"
-#include "parser.h"
-
-Node* parse_primary_expression(Parser* parser);
-
-#endif
