--- w0/parse_statement.h
@@ -1,9 +1,0 @@
-#ifndef WHIST_PARSE_STATEMENT_H
-#define WHIST_PARSE_STATEMENT_H
-
-#include "ast.h"
-#include "parser.h"
-
-Node* parse_statement(Parser* parser);
-
-#endif
