--- w0/parse_struct_init.h
@@ -1,9 +1,0 @@
-#ifndef WHIST_PARSE_STRUCT_INIT_H
-#define WHIST_PARSE_STRUCT_INIT_H
-
-#include "ast.h"
-#include "parser.h"
-
-Node* parse_struct_init(Parser* parser);
-
-#endif
