--- w0/parse_type.c
@@ -1,89 +1,0 @@
-#include "parse_expression.h"
-#include "parser_util.h"
-
-// Type parsing
-Node* parse_type(Parser* parser) {
-    Token token = parser->current;
-
-    // Pointer types are no longer supported - error if we see *
-    if (check_token(parser, TOK_STAR)) {
-        parse_error(parser, "Pointer types (*T) are no longer supported; use struct references");
-        return NULL;
-    }
-
-    // Tuple type: (T1, T2, ...)
-    if (match_token(parser, TOK_LPAREN)) {
-        Node* node = node_new(NODE_TUPLE_TYPE, token.line, token.column);
-        if (!node) {
-            parse_error(parser, "Out of memory");
-            return NULL;
-        }
-        nodelist_init(&node->as.tuple_type.elem_types);
-
-        // Parse first type
-        Node* first_type = parse_type(parser);
-        if (!first_type) {
-            node_free(node);
-            return NULL;
-        }
-        nodelist_push(&node->as.tuple_type.elem_types, first_type);
-
-        // Require at least one comma (i.e., at least 2 elements for a tuple)
-        if (!check_token(parser, TOK_COMMA)) {
-            parse_error(parser, "Tuple type requires at least 2 elements");
-            node_free(node);
-            return NULL;
-        }
-
-        // Parse remaining types
-        while (match_token(parser, TOK_COMMA)) {
-            Node* elem_type = parse_type(parser);
-            if (!elem_type) {
-                node_free(node);
-                return NULL;
-            }
-            nodelist_push(&node->as.tuple_type.elem_types, elem_type);
-        }
-
-        consume_token(parser, TOK_RPAREN, "Expected ')' after tuple type");
-        return node;
-    }
-
-    // Array type [n]type
-    if (match_token(parser, TOK_LBRACKET)) {
-        Node* size = NULL;
-        if (!check_token(parser, TOK_RBRACKET)) {
-            size = parse_expression(parser);
-        }
-        consume_token(parser, TOK_RBRACKET, "Expected ']' in array type");
-        Node* elem = parse_type(parser);
-
-        Node* node = node_new(NODE_INDEX, token.line, token.column);
-        if (!node) {
-            parse_error(parser, "Out of memory");
-            return NULL;
-        }
-        node->as.index.object = elem;
-        node->as.index.index  = size;
-        return node;
-    }
-
-    // Named type
-    if (match_token(parser, TOK_IDENT)) {
-        Node* node = node_new(NODE_IDENT, token.line, token.column);
-        if (!node) {
-            parse_error(parser, "Out of memory");
-            return NULL;
-        }
-        node->as.ident.name = copy_token_string(&token);
-        if (!node->as.ident.name) {
-            parse_error(parser, "Out of memory");
-            return NULL;
-        }
-        node->as.ident.length = token.length;
-        return node;
-    }
-
-    parse_error(parser, "Expected type");
-    return NULL;
-}
