--- w0/parse_type.h
@@ -1,9 +1,0 @@
-#ifndef WHIST_PARSE_TYPE_H
-#define WHIST_PARSE_TYPE_H
-
-#include "ast.h"
-#include "parser.h"
-
-Node* parse_type(Parser* parser);
-
-#endif
