--- w0/parse_var_decl.h
@@ -1,9 +1,0 @@
-#ifndef WHIST_PARSE_VAR_DECL_H
-#define WHIST_PARSE_VAR_DECL_H
-
-#include "ast.h"
-#include "parser.h"
-
-Node* parse_var_decl(Parser* parser, int is_const, int is_public);
-
-#endif
