--- w0/parse_while_stmt.h
@@ -1,9 +1,0 @@
-#ifndef WHIST_PARSE_WHILE_STMT_H
-#define WHIST_PARSE_WHILE_STMT_H
-
-#include "ast.h"
-#include "parser.h"
-
-Node* parse_while_stmt(Parser* parser);
-
-#endif
