--- w0/parser_util.c
@@ -1,106 +1,0 @@
-#include "parser_util.h"
-
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-
-void advance_token(Parser* parser) {
-    parser->previous = parser->current;
-
-    for (;;) {
-        parser->current = lexer_next(&parser->lexer);
-        if (parser->current.type != TOK_ERROR)
-            break;
-
-        // Report lexer error - the start field contains the error message
-        if (!parser->panic_mode) {
-            parser->panic_mode = 1;
-            parser->had_error  = 1;
-            fprintf(stderr, "[line %d:%d] Error: %.*s\n", parser->current.line,
-                    parser->current.column, (int)parser->current.length, parser->current.start);
-        }
-    }
-}
-
-int check_token(Parser* parser, TokenType type) {
-    return parser->current.type == type;
-}
-
-int match_token(Parser* parser, TokenType type) {
-    if (!check_token(parser, type))
-        return 0;
-    advance_token(parser);
-    return 1;
-}
-
-void parse_error_at(Parser* parser, Token* token, const char* message) {
-    if (parser->panic_mode)
-        return;
-    parser->panic_mode = 1;
-    parser->had_error  = 1;
-
-    if (token->type == TOK_EOF) {
-        snprintf(parser->error_msg, sizeof(parser->error_msg), "[line %d:%d] Error at end: %s",
-                 token->line, token->column, message);
-    } else if (token->type == TOK_ERROR) {
-        snprintf(parser->error_msg, sizeof(parser->error_msg), "[line %d:%d] Error: %s",
-                 token->line, token->column, message);
-    } else {
-        snprintf(parser->error_msg, sizeof(parser->error_msg), "[line %d:%d] Error at '%.*s': %s",
-                 token->line, token->column, (int)token->length, token->start, message);
-    }
-
-    fprintf(stderr, "%s\n", parser->error_msg);
-}
-
-void parse_error(Parser* parser, const char* message) {
-    parse_error_at(parser, &parser->current, message);
-}
-
-void consume_token(Parser* parser, TokenType type, const char* message) {
-    if (parser->current.type == type) {
-        advance_token(parser);
-        return;
-    }
-    parse_error(parser, message);
-}
-
-void synchronize(Parser* parser) {
-    parser->panic_mode = 0;
-
-    // Always advance at least once to ensure progress and prevent infinite loops
-    if (parser->current.type != TOK_EOF) {
-        advance_token(parser);
-    }
-
-    while (parser->current.type != TOK_EOF) {
-        if (parser->previous.type == TOK_SEMICOLON)
-            return;
-
-        switch (parser->current.type) {
-        case TOK_FUNC:
-        case TOK_STRUCT:
-        case TOK_ENUM:
-        case TOK_VAR:
-        case TOK_CONST:
-        case TOK_IF:
-        case TOK_WHILE:
-        case TOK_FOR:
-        case TOK_RETURN:
-            return;
-        default:
-            break;
-        }
-        advance_token(parser);
-    }
-}
-
-char* copy_token_string(Token* token) {
-    char* str = malloc(token->length + 1);
-    if (!str) {
-        return NULL;
-    }
-    memcpy(str, token->start, token->length);
-    str[token->length] = '\0';
-    return str;
-}
