--- w0/parser_util.h
@@ -1,22 +1,0 @@
-#ifndef WHIST_PARSER_UTIL_H
-#define WHIST_PARSER_UTIL_H
-
-#include "lexer.h"
-#include "parser.h"
-
-// Maximum recursion depth to prevent stack overflow
-#define MAX_PARSE_DEPTH 256
-
-// Current recursion depth for expression parsing
-extern int parse_depth;
-
-void  advance_token(Parser* parser);
-int   check_token(Parser* parser, TokenType type);
-int   match_token(Parser* parser, TokenType type);
-void  parse_error_at(Parser* parser, Token* token, const char* message);
-void  parse_error(Parser* parser, const char* message);
-void  consume_token(Parser* parser, TokenType type, const char* message);
-void  synchronize(Parser* parser);
-char* copy_token_string(Token* token);
-
-#endif
